+++ iblrig/gui/wizard.py
@@ -7,6 +7,8 @@ from pathlib import Path
 import shutil
 import subprocess
 import sys
+from typing import Any, Callable, Union, Optional, Iterable
+
 import yaml
 import traceback
 import webbrowser
@@ -14,12 +16,17 @@ import ctypes
 import os
 
 from PyQt5 import QtWidgets, QtCore, QtGui
-from PyQt5.QtCore import QThread
+from PyQt5.QtCore import QThread, QThreadPool
 from PyQt5.QtWidgets import QStyle
 
 from one.api import ONE
 import iblrig_tasks
-import iblrig_custom_tasks
+try:
+    import iblrig_custom_tasks
+    CUSTOM_TASKS = True
+except ImportError:
+    CUSTOM_TASKS = False
+    pass
 import iblrig.path_helper
 from iblrig.constants import BASE_DIR
 from iblrig.misc import _get_task_argument_parser
@@ -29,7 +36,10 @@ from iblrig.version_management import check_for_updates, get_changelog, is_dirty
 from iblrig.gui.ui_wizard import Ui_wizard
 from iblrig.gui.ui_update import Ui_update
 from iblrig.choiceworld import get_subject_training_info
+from iblutil.util import setup_logger
 from pybpodapi import exceptions
+
+log = setup_logger("iblrig")
 
 PROCEDURES = [
     'Behavior training/tasks',
@@ -45,7 +55,11 @@ PROJECTS = [
     'practice'
 ]
 
-WIZARD_PNG = str(Path(BASE_DIR).joinpath('iblrig', 'gui', 'wizard.png'))
+GUI_DIR = Path(BASE_DIR).joinpath('iblrig', 'gui')
+WIZARD_PNG = str(GUI_DIR.joinpath('wizard.png'))
+ICON_FLUSH = str(GUI_DIR.joinpath('icon_flush.svg'))
+ICON_HELP = str(GUI_DIR.joinpath('icon_help.svg'))
+ICON_STATUS_LED = str(GUI_DIR.joinpath('icon_status_led.svg'))
 
 
 # this class gets called to get the path constructor utility to predict the session path
@@ -59,38 +73,42 @@ class EmptySession(BaseSession):
         pass
 
 
-def _set_list_view_from_string_list(uilist: QtWidgets.QListView, string_list: list):
+def _set_list_view_from_string_list(ui_list: QtWidgets.QListView, string_list: list):
     """Small boiler plate util to set the selection of a list view from a list of strings"""
     if string_list is None or len(string_list) == 0:
         return
-    for i, s in enumerate(uilist.model().stringList()):
+    for i, s in enumerate(ui_list.model().stringList()):
         if s in string_list:
-            uilist.selectionModel().select(uilist.model().createIndex(i, 0), QtCore.QItemSelectionModel.Select)
+            ui_list.selectionModel().select(ui_list.model().createIndex(i, 0), QtCore.QItemSelectionModel.Select)
 
 
 @dataclass
 class RigWizardModel:
-    one: ONE = None
-    procedures: list = None
-    projects: list = None
-    task_name: str = None
-    user: str = None
-    subject: str = None
-    session_folder: Path = None
-    hardware_settings: dict = None
-    test_subject_name: str = 'test_subject'
+    one: Optional[ONE] = None
+    procedures: Optional[list] = None
+    projects: Optional[list] = None
+    task_name: Optional[str] = None
+    user: Optional[str] = None
+    subject: Optional[str] = None
+    session_folder: Optional[Path] = None
+    hardware_settings: Optional[dict] = None
+    test_subject_name: Optional[str] = 'test_subject'
     subject_details_worker = None
-    subject_details: tuple = None
+    subject_details: Optional[tuple] = None
 
     def __post_init__(self):
         self.iblrig_settings = iblrig.path_helper.load_settings_yaml()
         self.all_users = [self.iblrig_settings['ALYX_USER']] if self.iblrig_settings['ALYX_USER'] else []
         self.all_procedures = sorted(PROCEDURES)
-        # for the tasks, we build a dictionary that contains the task name as key and the path to the task.py as value
+
+        # for the tasks, we build a dictionary that contains the task name as key and the path to task.py as value
         tasks = sorted([p for p in Path(iblrig_tasks.__file__).parent.rglob('task.py')])
-        tasks.extend(sorted([p for p in Path(iblrig_custom_tasks.__file__).parent.rglob('task.py')]))
+        if CUSTOM_TASKS:
+            tasks.extend(sorted([p for p in Path(iblrig_custom_tasks.__file__).parent.rglob('task.py')]))
         self.all_tasks = OrderedDict({p.parts[-2]: p for p in tasks})
+
         self.all_projects = sorted(PROJECTS)
+
         # get the subjects from iterating over folders in the the iblrig data path
         if self.iblrig_settings['iblrig_local_data_path'] is None:
             self.all_subjects = [self.test_subject_name]
@@ -149,9 +167,15 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
         self.setupUi(self)
         self.setWindowIcon(QtGui.QIcon(WIZARD_PNG))
 
-        self.settings = QtCore.QSettings('iblrig', 'wizard')
+        self.settings = QtCore.QSettings()
+        self.move(self.settings.value("pos", self.pos(), QtCore.QPoint))
+
         self.model = RigWizardModel()
         self.model2view()
+
+        # default to biasedChoiceWorld
+        if (idx := self.uiComboTask.findText('_iblrig_tasks_biasedChoiceWorld')) >= 0:
+            self.uiComboTask.setCurrentIndex(idx)
 
         # connect widgets signals to slots
         self.uiComboTask.currentTextChanged.connect(self.controls_for_extra_parameters)
@@ -165,16 +189,20 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
         self.uiPushConnect.clicked.connect(self.alyx_connect)
         self.lineEditSubject.textChanged.connect(self._filter_subjects)
 
+        self.uiPushStatusLED.setChecked(self.settings.value("bpod_status_led", True, bool))
+        self.uiPushStatusLED.toggled.connect(self.toggle_status_led)
+        self.toggle_status_led(self.uiPushStatusLED.isChecked())
+
         self.running_task_process = None
         self.task_arguments = dict()
         self.task_settings_widgets = None
 
-        self.uiPushStart.setIcon(self.style().standardIcon(QStyle.SP_MediaPlay))
         self.uiPushStart.installEventFilter(self)
-
+        self.uiPushStart.setIcon(self.style().standardIcon(QStyle.SP_MediaPlay))
         self.uiPushPause.setIcon(self.style().standardIcon(QStyle.SP_MediaPause))
-        self.uiPushFlush.setIcon(self.style().standardIcon(QStyle.SP_BrowserReload))
-        self.uiPushHelp.setIcon(self.style().standardIcon(QStyle.SP_DialogHelpButton))
+        self.uiPushFlush.setIcon(QtGui.QIcon(ICON_FLUSH))
+        self.uiPushHelp.setIcon(QtGui.QIcon(ICON_HELP))
+        self.uiPushStatusLED.setIcon(QtGui.QIcon(ICON_STATUS_LED))
 
         self.controller2model()
 
@@ -188,17 +216,72 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
         total_space, total_used, total_free = shutil.disk_usage(local_data.anchor)
         self.uiProgressDiskSpace.setStatusTip(f'utilization of drive {local_data.anchor}')
         self.uiProgressDiskSpace.setValue(round(total_used / total_space * 100))
-        self.uiLableDiskAvailableValue.setText(f'{total_free / 1024 ** 3 : .1f} GB')
-        self.uiLableDiskIblrigValue.setText(f'{v8data_size / 1024 ** 3 : .1f} GB')
+        self.uiLabelDiskAvailableValue.setText(f'{total_free / 1024 ** 3 : .1f} GB')
+        self.uiLabelDiskIblrigValue.setText(f'{v8data_size / 1024 ** 3 : .1f} GB')
 
         tmp = QtWidgets.QLabel(f'iblrig v{iblrig.__version__}')
         tmp.setContentsMargins(4, 0, 0, 0)
         self.statusbar.addWidget(tmp)
         self.controls_for_extra_parameters()
 
-        self.update_check = UpdateCheckWorker(self)
+        self.layout().setSizeConstraint(QtWidgets.QLayout.SetFixedSize)
+        self.setWindowFlags(self.windowFlags() & ~QtCore.Qt.WindowFullscreenButtonHint)
 
-        QtCore.QTimer.singleShot(100, self.check_dirty)
+        # get AnyDesk ID
+        # anydesk_worker = Worker(get_anydesk_id)
+        # anydesk_worker.signals.result.connect(lambda var: print(f'Your AnyDesk ID: {var:s}'))
+        # QThreadPool.globalInstance().tryStart(anydesk_worker)
+
+        # check for update
+        update_worker = Worker(check_for_updates)
+        update_worker.signals.result.connect(self._on_check_update_result)
+        QThreadPool.globalInstance().start(update_worker)
+
+        # check dirty state
+        dirty_worker = Worker(is_dirty)
+        dirty_worker.signals.result.connect(self._on_check_dirty_result)
+        QThreadPool.globalInstance().start(dirty_worker)
+
+    def _on_check_update_result(self, result: tuple[bool, str]) -> None:
+        """
+        Handle the result of checking for updates.
+
+        Parameters
+        ----------
+        result : tuple[bool, str | None]
+            A tuple containing a boolean flag indicating update availability (result[0])
+            and the remote version string (result[1]).
+
+        Returns
+        -------
+        None
+        """
+        if result[0]:
+            UpdateNotice(parent=self, version=result[1])
+
+    def _on_check_dirty_result(self, repository_is_dirty: bool) -> None:
+        """
+        Handle the result of checking for local changes in the repository.
+
+        Parameters
+        ----------
+        repository_is_dirty : bool
+            A boolean flag indicating whether the repository contains local changes.
+
+        Returns
+        -------
+        None
+        """
+        if repository_is_dirty:
+            msg_box = QtWidgets.QMessageBox(parent=self)
+            msg_box.setWindowTitle('Warning')
+            msg_box.setIcon(QtWidgets.QMessageBox().Warning)
+            msg_box.setText("Your copy of iblrig contains local changes.\nDon't expect things to work as intended!")
+            msg_box.setDetailedText("To list all files that have been changed locally:\n\n"
+                                    "    git diff --name-only\n\n"
+                                    "To reset the repository to its default state:\n\n"
+                                    "    git reset --hard")
+            msg_box.exec()
 
     def eventFilter(self, obj, event):
         if obj == self.uiPushStart and event.type() in [QtCore.QEvent.HoverEnter, QtCore.QEvent.HoverLeave]:
@@ -213,47 +296,31 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
             return True
         return False
 
-    def closeEvent(self, event):
-        if self.running_task_process is None:
+    def closeEvent(self, event) -> None:
+        def accept() -> None:
+            self.settings.setValue("pos", self.pos())
+            self.settings.setValue("bpod_status_led", self.uiPushStatusLED.isChecked())
+            self.toggle_status_led(is_toggled=True)
+            bpod = Bpod(self.model.hardware_settings['device_bpod']['COM_BPOD'])  # bpod is a singleton
+            bpod.close()
             event.accept()
+
+        if self.running_task_process is None:
+            accept()
         else:
-            msgBox = QtWidgets.QMessageBox(parent=self)
-            msgBox.setWindowTitle("Hold on")
-            msgBox.setText("A task is running - do you really want to quit?")
-            msgBox.setStandardButtons(QtWidgets.QMessageBox.No | QtWidgets.QMessageBox.Yes)
-            msgBox.setIcon(QtWidgets.QMessageBox().Question)
-            match msgBox.exec_():
+            msg_box = QtWidgets.QMessageBox(parent=self)
+            msg_box.setWindowTitle("Hold on")
+            msg_box.setText("A task is running - do you really want to quit?")
+            msg_box.setStandardButtons(QtWidgets.QMessageBox.No | QtWidgets.QMessageBox.Yes)
+            msg_box.setIcon(QtWidgets.QMessageBox().Question)
+            match msg_box.exec_():
                 case QtWidgets.QMessageBox.No:
                     event.ignore()
                 case QtWidgets.QMessageBox.Yes:
                     self.setEnabled(False)
                     self.repaint()
                     self.start_stop()
-                    event.accept()
-
-    def check_dirty(self):
-        """
-        Check if the iblrig installation contains local changes.
-
-        This method checks if the installed version of iblrig contains local changes
-        (indicated by the version string ending with 'dirty'). If local changes are
-        detected, it displays a warning message to inform the user about potential
-        issues and provides instructions on how to reset the repository to its
-        default state.
-
-        Returns
-        -------
-        None
-        """
-        if not is_dirty():
-            return
-        msg_box = QtWidgets.QMessageBox(parent=self)
-        msg_box.setWindowTitle("Warning")
-        msg_box.setText("Your copy of iblrig contains local changes.\nDon't expect things to work as intended!")
-        msg_box.setDetailedText("To reset the repository to its default state, use:\n\n    git reset --hard\n\n")
-        msg_box.setStandardButtons(QtWidgets.QMessageBox.Ok)
-        msg_box.setIcon(QtWidgets.QMessageBox().Information)
-        msg_box.exec_()
+                    accept()
 
     def model2view(self):
         # stores the current values in the model
@@ -308,7 +375,7 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
                 widget.setTristate(False)
                 if arg.default:
                     widget.setCheckState(arg.default * 2)
-                widget.toggled.connect(lambda val, p=param: self._set_task_arg(param, val > 0))
+                widget.toggled.connect(lambda val, p=param: self._set_task_arg(p, val > 0))
                 widget.toggled.emit(widget.isChecked() > 0)
 
             # create widget for string arguments
@@ -329,6 +396,16 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
                     widget.editingFinished.connect(
                         lambda p=param, w=widget: self._set_task_arg(p, w.text()))
                     widget.editingFinished.emit()
+
+            # create widget for list of floats
+            elif arg.type == float and arg.nargs == '+':
+                widget = QtWidgets.QLineEdit()
+                if arg.default:
+                    widget.setText(str(arg.default)[1:-1])
+                widget.editingFinished.connect(
+                    lambda p=param, w=widget:
+                    self._set_task_arg(p, [x.strip() for x in w.text().split(',')]))
+                widget.editingFinished.emit()
 
             # create widget for numerical arguments
             elif arg.type in [float, int]:
@@ -358,6 +435,15 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
 
             # some customizations
             match widget.property('parameter_dest'):
+                case 'probability_left':
+                    widget.setMinimum(0.0)
+                    widget.setMaximum(1.0)
+                    widget.setSingleStep(0.1)
+                    widget.setDecimals(2)
+
+                case 'contrast_set_probability_type':
+                    label = 'Probability Type'
+
                 case 'session_template_id':
                     label = 'Session Template ID'
 
@@ -382,6 +468,26 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
                         self._set_task_arg(a.option_strings[0], str(val if val > m else -1)))
                     widget.valueChanged.emit(widget.value())
 
+                case 'adaptive_gain':
+                    label = 'Stimulus Gain, μl'
+                    widget.setSpecialValueText('automatic')
+                    widget.setMaximum(3)
+                    widget.setSingleStep(0.1)
+                    widget.setMinimum(1.4)
+                    widget.setValue(widget.minimum())
+                    widget.valueChanged.connect(
+                        lambda val, a=arg, m=widget.minimum():
+                        self._set_task_arg(a.option_strings[0], str(val if val > m else -1)))
+                    widget.valueChanged.emit(widget.value())
+
+                case 'reward_amount_ul':
+                    label = 'Reward Amount, μl'
+                    widget.setSingleStep(0.1)
+                    widget.setMinimum(0)
+
+                case 'stim_gain':
+                    label = 'Stimulus Gain'
+
             layout.addRow(self.tr(label), widget)
 
         # add label to indicate absence of task specific parameters
@@ -389,14 +495,11 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
             layout.addRow(self.tr('(none)'), None)
             layout.itemAt(0, 0).widget().setEnabled(False)
 
-        # call timer to set size of window
-        QtCore.QTimer.singleShot(1, self.set_size)
+        # call timer to resize window
+        QtCore.QTimer.singleShot(1, lambda: self.resize(self.minimumSizeHint()))
 
     def _set_task_arg(self, key, value):
         self.task_arguments[key] = value
-
-    def set_size(self):
-        self.setFixedSize(self.layout().minimumSize())
 
     def alyx_connect(self):
         self.model.connect()
@@ -444,6 +547,13 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
                     self.model.session_folder.joinpath('.stop').unlink()
                 self.model.raw_data_folder = task.paths['SESSION_RAW_DATA_FOLDER']
 
+                # disable Bpod status LED
+                bpod = Bpod(self.model.hardware_settings['device_bpod']['COM_BPOD'])
+                bpod.set_status_led(False)
+
+                # close Bpod singleton so subprocess can access use the port
+                bpod.close()
+
                 # runs the python command
                 cmd = [shutil.which('python')]
                 if self.model.task_name:
@@ -457,12 +567,19 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
                 if self.model.projects:
                     cmd.extend(['--projects', *self.model.projects])
                 for key in self.task_arguments.keys():
-                    cmd.extend([key, self.task_arguments[key]])
+                    if isinstance(self.task_arguments[key], Iterable) and not isinstance(self.task_arguments[key], str):
+                        cmd.extend([str(key)])
+                        for value in self.task_arguments[key]:
+                            cmd.extend([value])
+                    else:
+                        cmd.extend([key, self.task_arguments[key]])
                 cmd.extend(['--weight', f'{weight}'])
                 cmd.append('--wizard')
                 if self.uiCheckAppend.isChecked():
                     cmd.append('--append')
                 if self.running_task_process is None:
+                    log.info('Starting subprocess')
+                    log.info(subprocess.list2cmdline(cmd))
                     # self.running_task_process = QProcess()
                     # self.running_task_process.start(shutil.which('python'), cmd)
                     # self.running_task_process.readyReadStandardOutput.connect(self.handle_stdout)
@@ -491,7 +608,11 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
                 self.uiPushStart.setText('Start')
                 self.uiPushStart.setStatusTip('start the session')
                 self.uiPushStart.setIcon(self.style().standardIcon(QStyle.SP_MediaPlay))
+
+                # re-enable UI elements and recall state of Bpod status LED
                 self.enable_UI_elements()
+                bpod = Bpod(self.model.hardware_settings['device_bpod']['COM_BPOD'])
+                bpod.set_status_led(self.uiPushStatusLED.isChecked())
 
                 if (task_settings_file := Path(self.model.raw_data_folder).joinpath("_iblrig_taskSettings.raw.json")).exists():
                     with open(task_settings_file, "r") as fid:
@@ -539,8 +660,19 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
             self.uiPushFlush.setStyleSheet('')
             return
 
-        if not self.uiPushFlush.isChecked():
-            bpod.close()
+    def toggle_status_led(self, is_toggled: bool):
+
+        # paint button green when in toggled state
+        self.uiPushStatusLED.setStyleSheet('QPushButton {background-color: rgb(128, 255, 128);}'
+                                           if is_toggled else '')
+        self.enable_UI_elements()
+
+        try:
+            bpod = Bpod(self.model.hardware_settings['device_bpod']['COM_BPOD'])
+            bpod.set_status_led(is_toggled)
+        except (OSError, exceptions.bpod_error.BpodErrorException, AttributeError):
+            self.uiPushStatusLED.setChecked(False)
+            self.uiPushStatusLED.setStyleSheet('')
 
     def help(self):
         webbrowser.open('https://int-brain-lab.github.io/iblrig/usage.html')
@@ -554,6 +686,7 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
             and len(self.uiListProcedures.selectedIndexes()) > 0)
         self.uiPushPause.setEnabled(is_running)
         self.uiPushFlush.setEnabled(not is_running)
+        self.uiPushStatusLED.setEnabled(not is_running)
         self.uiCheckAppend.setEnabled(not is_running)
         self.uiGroupParameters.setEnabled(not is_running)
         self.uiGroupTaskParameters.setEnabled(not is_running)
@@ -561,40 +694,155 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
         self.repaint()
 
 
-class UpdateCheckWorker(QThread):
+class WorkerSignals(QtCore.QObject):
+    """
+    Signals used by the Worker class to communicate with the main thread.
 
-    def __init__(self, parent):
-        super().__init__()
-        self.parent = parent
-        self.update_available = False
-        self.remote_version = ''
-        self.finished.connect(self.check_results)
-        self.start()
+    Attributes
+    ----------
+    finished : QtCore.pyqtSignal
+        Signal emitted when the worker has finished its task.
 
-    def run(self):
-        self.update_available, self.remote_version = check_for_updates()
+    error : QtCore.pyqtSignal(tuple)
+        Signal emitted when an error occurs. The signal carries a tuple with the exception type,
+        exception value, and the formatted traceback.
 
-    def check_results(self):
-        if self.update_available:
-            dialog = self.UpdateNotice(parent=self.parent)
-            dialog.uiLabelHeader.setText(f"Update to iblrig {self.remote_version} is available.")
-            dialog.uiPushButtonOK.released.connect(lambda: dialog.close())
-            dialog.exec_()
-            self.deleteLater()
+    result : QtCore.pyqtSignal(Any)
+        Signal emitted when the worker has successfully completed its task. The signal carries
+        the result of the task.
 
-    class UpdateNotice(QtWidgets.QDialog, Ui_update):
-        def __init__(self, parent=None):
-            super().__init__(parent)
-            self.setupUi(self)
-            self.uiTextBrowserChanges.setMarkdown(get_changelog())
-            self.uiLabelLogo.setPixmap(QtGui.QPixmap(WIZARD_PNG))
-            self.setWindowIcon(QtGui.QIcon(WIZARD_PNG))
-            self.setWindowFlags(self.windowFlags() & ~QtCore.Qt.WindowContextHelpButtonHint)
+    progress : QtCore.pyqtSignal(int)
+        Signal emitted to report progress during the task. The signal carries an integer value.
+    """
+    finished: QtCore.pyqtSignal = QtCore.pyqtSignal()
+    error: QtCore.pyqtSignal = QtCore.pyqtSignal(tuple)
+    result: QtCore.pyqtSignal = QtCore.pyqtSignal(object)
+    progress: QtCore.pyqtSignal = QtCore.pyqtSignal(int)
+
+
+class Worker(QtCore.QRunnable):
+    """
+    A generic worker class for executing functions concurrently in a separate thread.
+
+    This class is designed to run functions concurrently in a separate thread and emit signals
+    to communicate the results or errors back to the main thread.
+
+    Adapted from: https://www.pythonguis.com/tutorials/multithreading-pyqt-applications-qthreadpool/
+
+    Attributes
+    ----------
+    fn : Callable
+        The function to be executed concurrently.
+
+    args : tuple
+        Positional arguments for the function.
+
+    kwargs : dict
+        Keyword arguments for the function.
+
+    signals : WorkerSignals
+        An instance of WorkerSignals used to emit signals.
+
+    Methods
+    -------
+    run() -> None
+        The main entry point for running the worker. Executes the provided function and
+        emits signals accordingly.
+    """
+
+    def __init__(self, fn: Callable[..., Any], *args: Any, **kwargs: Any) -> None:
+        """
+        Initialize the Worker instance.
+
+        Parameters
+        ----------
+        fn : Callable
+            The function to be executed concurrently.
+
+        *args : tuple
+            Positional arguments for the function.
+
+        **kwargs : dict
+            Keyword arguments for the function.
+
+        Returns
+        -------
+        None
+        """
+        super(Worker, self).__init__()
+        self.fn: Callable[..., Any] = fn
+        self.args = args
+        self.kwargs = kwargs
+        self.signals: WorkerSignals = WorkerSignals()
+
+    def run(self) -> None:
+        """
+        Execute the provided function and emit signals accordingly.
+
+        This method is the main entry point for running the worker. It executes the provided
+        function and emits signals to communicate the results or errors back to the main thread.
+
+        Returns
+        -------
+        None
+        """
+        try:
+            result = self.fn(*self.args, **self.kwargs)
+        except:  # noqa: 722
+            # Handle exceptions and emit error signal with exception details
+            traceback.print_exc()
+            exctype, value = sys.exc_info()[:2]
+            self.signals.error.emit((exctype, value, traceback.format_exc()))
+        else:
+            # Emit result signal with the result of the task
+            self.signals.result.emit(result)
+        finally:
+            # Emit the finished signal to indicate completion
+            self.signals.finished.emit()
+
+
+class UpdateNotice(QtWidgets.QDialog, Ui_update):
+    """
+    A dialog for displaying update notices.
+
+    This class is used to create a dialog for displaying update notices.
+    It shows information about the available update and provides a changelog.
+
+    Parameters
+    ----------
+    parent : QtWidgets.QWidget
+        The parent widget associated with this dialog.
+
+    update_available : bool
+        Indicates if an update is available.
+
+    version : str
+        The version of the available update.
+
+    Attributes
+    ----------
+    None
+
+    Methods
+    -------
+    None
+    """
+
+    def __init__(self, parent: QtWidgets.QWidget, version: str) -> None:
+        super().__init__(parent)
+        self.setupUi(self)
+        self.setAttribute(QtCore.Qt.WA_DeleteOnClose, True)
+        self.setWindowIcon(QtGui.QIcon(WIZARD_PNG))
+        self.uiLabelLogo.setPixmap(QtGui.QPixmap(WIZARD_PNG))
+        self.uiLabelHeader.setText(f"Update to iblrig {version} is available.")
+        self.uiTextBrowserChanges.setMarkdown(get_changelog())
+        self.setWindowFlags(self.windowFlags() & ~QtCore.Qt.WindowContextHelpButtonHint)
+        self.exec()
 
 
 class SubjectDetailsWorker(QThread):
-    subject_name: str = None
-    result: tuple[int, float, dict] = None
+    subject_name: Union[str, None] = None
+    result: Union[tuple[dict, dict], None] = None
 
     def __init__(self, subject_name):
         super().__init__()
@@ -605,9 +853,13 @@ class SubjectDetailsWorker(QThread):
 
 
 def main():
+    QtCore.QCoreApplication.setOrganizationName("International Brain Laboratory")
+    QtCore.QCoreApplication.setOrganizationDomain("internationalbrainlab.org")
+    QtCore.QCoreApplication.setApplicationName("IBLRIG Wizard")
+
     if os.name == 'nt':
-        appid = f'IBL.iblrig.wizard.{iblrig.__version__}'
-        ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(appid)
+        app_id = f'IBL.iblrig.wizard.{iblrig.__version__}'
+        ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(app_id)
     app = QtWidgets.QApplication(sys.argv)
     app.setStyle("Fusion")
     w = RigWizard()